+++ CHANGELOG.md
@@ -24,6 +24,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 - Added Page component
 - Added PageLoader component
 - Added PageLoader component to PersistGate as loading state
+- Added Touchable component
 
 ### Removed